requirements.txt
black==22.10.0
click==8.1.3
joblib==1.2.0
mypy-extensions==0.4.3
numpy>=1.23.4
pandas>=1.5.1
pathspec==0.10.2
platformdirs==2.5.4
python-dateutil==2.8.2
pytz==2022.6
PyYAML>=6.0
scikit-learn>=1.1.3
scipy>=1.9.3
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
